DSMP/requirements.txt
python=3.11
Flask
google-genrativeai
dotenv
s